requirements.txt
openai==1.107.1
pandas==2.3.2
Pillow==11.3.0
python-dotenv==1.1.1
python_pptx==1.0.2
streamlit_agraph==0.0.45
streamlit_extras==0.7.8
langchain==0.3.27
langchain_community==0.3.29
langchain_openai==0.3.33
matplotlib==3.10.6
numpy==2.3.3
pdfplumber==0.11.7
psycopg2_binary==2.9.10
python-dotenv==1.1.1
python_pptx==1.0.2
sentence_transformers==5.1.0
streamlit==1.48.1
wget==3.2


# --- Database & drivers ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# --- Auth & security ---
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1

# --- Utilities ---
pydantic==2.7.1
requests==2.31.0
typing-extensions==4.11.0


